test t/02-snp-read.t: Added debug output
For some reason Perl 5.26.3 failed this test, but the very minor change to
display $@ caused the test to succeed.  Very strange...but just in case we now
report $@ so future debugging will be easier if this comes up again.

Also changed 3 MHz to 4 MHz so there are no duplicate frequencies.